@@ -6,32 +6,16 @@ use DateInterval;
 use FeatureTox\Model\Context;
 use Psr\Cache\CacheItemPoolInterface as CachePool;
 
-/**
- * Generic cache for activator
- *
- * @author Michel Chowanski <michel.chowanski@bestit-online.de>
- * @package FeatureTox\Activator
- */
 class CacheActivator implements FeatureActivatorInterface
 {
-    /**
-     * Cache key
-     */
-    const CACHE_KEY = 'FeatureTox';
+    public const CACHE_KEY = 'FeatureTox';
 
     /**
      * The origin activator
-     *
-     * @var FeatureActivatorInterface
      */
-    private $activator;
+    private FeatureActivatorInterface $activator;
 
-    /**
-     * The cache pool
-     *
-     * @var CachePool
-     */
-    private $cachePool;
+    private CachePool $cachePool;
 
     /**
      * Time to live for cache items
@@ -46,15 +30,8 @@ class CacheActivator implements FeatureActivatorInterface
      *
      * @var bool[]
      */
-    private $memory = [];
+    private array $memory = [];
 
-    /**
-     * CacheActivator constructor.
-     *
-     * @param FeatureActivatorInterface $activator
-     * @param CachePool $cachePool
-     * @param int|DateInterval|null $cacheTtl
-     */
     public function __construct(FeatureActivatorInterface $activator, CachePool $cachePool, $cacheTtl = 3600)
     {
         $this->activator = $activator;
@@ -62,28 +39,17 @@ class CacheActivator implements FeatureActivatorInterface
         $this->cacheTtl = $cacheTtl;
     }
 
-    /**
-     * Get activator
-     *
-     * @return FeatureActivatorInterface
-     */
-    public function getActivator()
+    public function getActivator(): FeatureActivatorInterface
     {
         return $this->activator;
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    public function getName()
+    public function getName(): string
     {
         return $this->activator->getName();
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    public function isActive($name, Context $context)
+    public function isActive($name, Context $context): bool
     {
         $hash = static::CACHE_KEY . '#' . $this->getName() . '#' . md5($name . '-' . $context->serialize());
 
@@ -94,20 +60,18 @@ class CacheActivator implements FeatureActivatorInterface
 
         // Step 2: Try get from (optional) cache
         $cacheItem = null;
-        if ($this->cachePool !== null) {
-            $cacheItem = $this->cachePool->getItem($hash);
+        $cacheItem = $this->cachePool->getItem($hash);
 
-            if ($cacheItem->isHit()) {
-                $this->memory[$hash] = $cacheItem->get();
-                return $this->memory[$hash];
-            }
+        if ($cacheItem->isHit()) {
+            $this->memory[$hash] = $cacheItem->get();
+            return $this->memory[$hash];
         }
 
         // Step 3: Get from activators and save to cache
         $this->memory[$hash] = $this->activator->isActive($name, $context);
 
         // Write result to cache
-        if ($this->cachePool !== null && $cacheItem !== null) {
+        if ($cacheItem !== null) {
             $cacheItem->set($this->memory[$hash]);
             $cacheItem->expiresAfter($this->cacheTtl);
             $this->cachePool->save($cacheItem);
